Re-split NO3 around e_breakable
