+++ src/st/no3/e_breakable.c
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "no3.h"
+#include "sfx.h"
+
+extern u16 g_eBreakableInit[];
+extern u8* g_eBreakableAnimations[8];
+extern u8 g_eBreakableHitboxes[];
+extern u8 g_eBreakableExplosionTypes[];
+extern u16 g_eBreakableanimSets[];
+extern u8 g_eBreakableDrawModes[];
+#include "../e_breakable.h"
