+++ src/st/no3/no3.h
@@ -46,7 +46,7 @@ typedef enum EntityIDs {
     /* 0x14 */ E_ID_14 = 0x14,
     /* 0x15 */ E_GREY_PUFF = 0x15,
     /* 0x17 */ E_BG_LIGHTNING = 0x17,
-    /* 0x1E */ E_CAVERN_DOOR_LEVER_UNK0 = 0x1E,
+    /* 0x1E */ E_CAVERN_DOOR_PLATFORM = 0x1E,
     /* 0x27 */ E_FALLING_ROCK_2 = 0x27,
     /* 0x29 */ E_LIGHTNING_THUNDER = 0x29,
     /* 0x2A */ E_LIGHTNING_CLOUD,
